AzureAgenticLab/requirements.txt
python-dotenv
requests
openai
azure-identity
azure-core
streamlit